+++ src/utils/assert.ts
@@ -11,12 +11,15 @@ const internalErrorPrefix = `[${npmPackage}][Internal Failure]`
 const usageErrorPrefix = `[${npmPackage}][Wrong Usage]`
 const warningPrefix = `[${npmPackage}][Warning]`
 
-function assert(condition: unknown): asserts condition {
+function assert(condition: unknown, debugInfo?: Record<string, unknown>): asserts condition {
   if (condition) {
     return
   }
+  const debugStr = !debugInfo
+    ? ''
+    : ` Debug info: \`${JSON.stringify(debugInfo)}\` (this if for the \`${libName}\` maintainers; you can ignore this).`
   const internalError = newError(
-    `${internalErrorPrefix} You stumbled upon an bug in \`${libName}\`'s source code (an internal assertion failed). This should definitely not be happening, and you should create a new issue at https://github.com/brillout/${libName}/issues/new that includes this error stack (the error stack is enough to debug this). Or reach out on Discord. A fix will be written promptly.`
+    `${internalErrorPrefix} You stumbled upon a bug in \`${libName}\`'s source code (an internal assertion failed). This should definitely not be happening, and you should create a new issue at https://github.com/brillout/${libName}/issues/new that includes this error stack (the error stack is enough to debug this). Or reach out on Discord. A fix will be written promptly.${debugStr}`
   )
   throw internalError
 }