+++ src/utils/throttle.ts
@@ -1,0 +1,14 @@
+export { throttle }
+
+function throttle(func: Function, waitTime: number) {
+  let isQueued: boolean = false
+  return () => {
+    if (!isQueued) {
+      isQueued = true
+      setTimeout(() => {
+        isQueued = false
+        func()
+      }, waitTime)
+    }
+  }
+}